fix(audit): code quality scan fixes
- fix(raft_log): initialize RaftLogEntry index_/term_ in default constructor
- fix(rpc): initialize RaftResponse type_ to KInvalid in default constructor
- fix(client_session): change ttl_ms from uint32_t to uint64_t to prevent overflow
- fix(election_manager): remove unused RequestVoteRequest/PreVoteRequest objects
- docs: add code-quality-scan-report-2025.md

All 325 tests pass (292 unit + 27 integration + 6 deterministic)

+++ include/rollingraft/client_session_manager.h
@@ -55,7 +55,7 @@ class ClientSessionManager {
    * @param ttl_ms Session TTL in milliseconds (0 = no expiration)
    */
   explicit ClientSessionManager(size_t max_sessions = 10000,
-                                uint32_t ttl_ms = 300000);
+                                uint64_t ttl_ms = 300000);
 
   /**
    * Check if a request is a duplicate and return cached result if so.
@@ -98,7 +98,7 @@ class ClientSessionManager {
  private:
   mutable std::mutex mtx_;
   size_t max_sessions_;
-  uint32_t ttl_ms_;
+  uint64_t ttl_ms_;
 
   // LRU list: front = most recently used, back = least recently used
   std::list<uint64_t> lru_list_;
